Refactor: Split Script into Multiple Modules
