+++ README.md
@@ -1,155 +1,116 @@
 # Style Transfer Visualizer
-![Python CI](https://github.com/bjg-gh/style_transfer_visualizer/actions/workflows/python-ci.yml/badge.svg)
-[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](LICENSE)
-[![codecov](https://codecov.io/gh/bjg-gh/style_transfer_visualizer/branch/ci-test/graph/badge.svg)](https://codecov.io/gh/bjg-gh/style_transfer_visualizer)
-![GitHub Release](https://img.shields.io/github/v/release/bjg-gh/style_transfer_visualizer?sort=semver)
-[![Release Notes](https://img.shields.io/badge/Release_Notes-📄%20View-blue)](./RELEASES.md)
 
-**Author**: [@bjg-gh](https://github.com/bjg-gh)
+A command-line tool that applies neural style transfer to images using PyTorch. It supports timelapse video generation, configuration via CLI or TOML, and flexible output control.
 
-A powerful and customizable neural style transfer tool implemented in PyTorch. This script allows you to stylize content images using the style of another image and generate a timelapse video and loss plot of the transfer process.
+## Features
 
-## 🚀 Features
+- Neural style transfer using VGG19
+- Configurable via CLI or TOML
+- Timelapse video generation with adjustable FPS and quality
+- Optional normalization and initialization methods
+- Save intermediate steps or final image only
+- Deterministic execution with `--seed`
+- Unit and integration tests with full coverage
 
-- Content and style image support using VGG19-based feature extraction
-- LBFGS optimizer with customizable initialization (random, white, content)
-- Frame-by-frame image saving and loss tracking
-- Optional timelapse MP4 video output and matplotlib loss plots
-- Flexible CLI with extensive configuration options
-- Comprehensive test suite with coverage and integration testing
+---
 
-## 🛠 Installation
+## Installation
+
+Create and activate a virtual environment:
 
 ```bash
 python -m venv venv
-venv\Scripts\activate  # Windows
+venv\Scripts\activate
+```
+
+Install dependencies:
+
+```bash
 pip install -r requirements.txt
 ```
 
-## 📸 Usage
+For CUDA support (recommended), use:
 
 ```bash
-python style_transfer_visualizer.py --content path/to/content.jpg --style path/to/style.jpg
+pip install -r requirements-cuda.txt
 ```
 
-### Example:
+---
+
+## Usage
+
+Run the tool from the command line:
 
 ```bash
-python style_transfer_visualizer.py --content cat.jpg --style starry_night.jpg --steps 500 --fps 30 --final-only
+python run_visualizer.py --content path/to/content.jpg --style path/to/style.jpg
 ```
 
-**Note:** Input images must be pre-sized by the user. Minimum dimension: 64px. Processing may be slow for images above 3000px.
+### Common Options
 
-### Options:
+- `--steps`, `--save-every`, `--style-w`, `--content-w`, `--lr`
+- `--init-method {random,white,content}`
+- `--fps`, `--quality`
+- `--no-normalize`, `--no-video`, `--final-only`
+- `--device cpu|cuda`
+- `--seed`
+- `--config config.toml`
 
-| Option              | Description                                                                 |
-|---------------------|-----------------------------------------------------------------------------|
-| `--content`         | Path to the content image (required)                                       |
-| `--style`           | Path to the style image (required)                                         |
-| `--output`          | Output directory (default: `out`)                                          |
-| `--steps`           | Number of optimization steps (default: 300)                                |
-| `--save-every`      | Save frame every N steps (default: 20)                                     |
-| `--style-w`         | Style loss weight (default: 1e6)                                           |
-| `--content-w`       | Content loss weight (default: 1.0)                                         |
-| `--lr`              | Learning rate for optimizer (default: 1.0)                                 |
-| `--fps`             | Frames per second for video (default: 10)                                  |
-| `--init-method`     | Initialization: `content`, `random`, or `white` (default: `random`)        |
-| `--no-normalize`    | Disable ImageNet normalization                                             |
-| `--no-video`        | Skip video generation                                                      |
-| `--final-only`      | Save only the final image (implies `--no-video`)                           |
-| `--quality`         | Video quality (1–10, default: 10)                                          |
-| `--device`          | Device to run on: `cpu` or `cuda` (default: `cuda`)                        |
-| `--seed`            | Random seed for reproducibility (default: 0)                               |
+---
 
-### 🗂 Optional: Using a Config File
+## Project Structure
 
-Instead of passing many CLI arguments, you can store settings in a `config.toml` file:
-
-```bash
-python style_transfer_visualizer.py --content cat.jpg --style mosaic.jpg --config path/to/config.toml
+```
+style_transfer_visualizer/
+├── run_visualizer.py              # Entry point
+├── __version__.py                 # Version string
+├── config.toml                    # Example config
+├── src/
+│   └── style_transfer_visualizer/
+│       ├── cli.py
+│       ├── config.py
+│       ├── config_defaults.py
+│       ├── constants.py
+│       ├── core_model.py
+│       ├── image_io.py
+│       ├── logging_utils.py
+│       ├── main.py
+│       ├── optimization.py
+│       ├── types.py
+│       ├── utils.py
+│       └── video.py
+├── tests/
+│   ├── test_cli.py
+│   ├── test_config.py
+│   ├── test_core_model.py
+│   ├── test_image_io.py
+│   ├── test_main.py
+│   ├── test_optimization.py
+│   ├── test_run_visualizer.py
+│   ├── test_utils.py
+│   ├── test_video.py
+│   └── conftest.py
 ```
 
-The config file supports the following sections:
+---
 
-```toml
-[output]
-output = "out/"
+## Testing
 
-[optimization]
-steps = 500
-style_w = 1e5
-content_w = 1.0
-lr = 1.0
-init_method = "random"
-seed = 42
-normalize = true
-
-[video]
-save_every = 10
-fps = 30
-quality = 9
-create_video = true
-final_only = false
-
-[hardware]
-device = "cuda"
-```
-
-- Any CLI argument will override its corresponding config value.
-- The config file is optional — default values still apply if it's omitted.
-
-## 📊 Output
-
-- `stylized_<content>_x_<style>.png`: Final stylized image
-- `timelapse_<content>_x_<style>.mp4`: Optional video output
-- `loss_plot.png`: Optional loss visualization
-
-## ✅ Testing
-
-This project uses `pytest` with coverage and visual/integration markers.
-
-Run all tests with coverage:
+Run all tests:
 
 ```bash
 pytest
 ```
 
-View HTML coverage report:
+Run with coverage:
 
 ```bash
-pytest --cov=. --cov-report=html
+pytest --cov=src/style_transfer_visualizer --cov-report=term --cov-report=html
 ```
 
-Run only fast unit tests:
+Open `htmlcov/index.html` in your browser for a visual coverage report.
 
-```bash
-pytest -m "not slow"
-```
+---
 
-## 🧪 Test Suite Highlights
+## License
 
-- `test_core_model.py`: VGG feature slicing, Gram matrix, forward loss computation
-- `test_cli.py`: Argument parsing, `main()` CLI entry
-- `test_image_processing.py`: Loading, resizing, padding, normalization
-- `test_utils.py`: Device setup, directory creation, seeding
-- `test_video_output.py`: Timelapse writing, final image output
-
-## 📁 Project Structure
-
-```
-.
-├── style_transfer_visualizer.py    # Main neural style transfer implementation
-├── requirements.txt                # Dependencies
-├── pytest.ini                      # Pytest configuration
-├── .coveragerc                     # Coverage configuration
-├── tests/
-    ├── test_cli.py
-    ├── test_core_model.py
-    ├── test_image_processing.py
-    ├── test_utils.py
-    └── test_video_output.py
-```
-
-## 📄 License
-
-MIT License. See `LICENSE` file for details.
+MIT License
