+++ pytest.ini
@@ -1,6 +1,7 @@
 # pytest.ini
 [pytest]
 testpaths = tests
+pythonpath = src
 python_files = test_*.py
 python_classes = Test*
 python_functions = test_*
