+++ src/style_transfer_visualizer/config_defaults.py
