+++ src/style_transfer_visualizer/constants.py
