+++ tests/test_config.py
@@ -8,16 +8,12 @@ Covers:
 - Structural validation of the merged configuration object
 """
 import tempfile
+
 import pytest
 import tomlkit
 from pydantic import ValidationError
-from config import (
-    VideoConfig,
-    OptimizationConfig,
-    ConfigLoader,
-    StyleTransferConfig
-)
-from config_defaults import (
+
+from style_transfer_visualizer.config_defaults import (
     DEFAULT_LEARNING_RATE,
     DEFAULT_FPS,
     DEFAULT_DEVICE,
@@ -29,6 +25,8 @@ from config_defaults import (
     DEFAULT_STEPS,
     DEFAULT_INIT_METHOD
 )
+import style_transfer_visualizer.config as stv_config
+
 
 def create_toml_file(data: dict) -> str:
     """Write a TOML string to a temporary file and return its path."""
@@ -55,9 +53,9 @@ def test_load_valid_config():
         "hardware": {"device": "cpu"}
     }
     path = create_toml_file(config_data)
-    cfg = ConfigLoader.load(path)
+    cfg = stv_config.ConfigLoader.load(path)
 
-    assert isinstance(cfg, StyleTransferConfig)
+    assert isinstance(cfg, stv_config.StyleTransferConfig)
     assert cfg.output.output == "results"
     assert cfg.optimization.steps == 500
     assert cfg.optimization.style_w == 123456.0
@@ -70,7 +68,7 @@ def test_load_valid_config():
 def test_missing_file_raises():
     """Ensure FileNotFoundError is raised for nonexistent config."""
     with pytest.raises(FileNotFoundError):
-        ConfigLoader.load("nonexistent_file.toml")
+        stv_config.ConfigLoader.load("nonexistent_file.toml")
 
 
 def test_partial_config_uses_defaults():
@@ -78,7 +76,7 @@ def test_partial_config_uses_defaults():
     path = create_toml_file({
         "optimization": {"steps": 42}
     })
-    cfg = ConfigLoader.load(path)
+    cfg = stv_config.ConfigLoader.load(path)
 
     assert cfg.optimization.steps == 42
     assert cfg.optimization.lr == DEFAULT_LEARNING_RATE
@@ -88,65 +86,69 @@ def test_partial_config_uses_defaults():
 def test_video_config_invalid_fps():
     """Ensure invalid fps raises ValidationError."""
     with pytest.raises(ValidationError) as exc_info:
-        VideoConfig(fps=0, quality=DEFAULT_VIDEO_QUALITY,
-                    save_every=DEFAULT_SAVE_EVERY)
+        stv_config.VideoConfig(fps=0, quality=DEFAULT_VIDEO_QUALITY,
+                               save_every=DEFAULT_SAVE_EVERY)
     assert "fps" in str(exc_info.value)
 
 
 def test_video_config_invalid_quality():
     """Ensure invalid quality raises ValidationError."""
     with pytest.raises(ValidationError) as exc_info:
-        VideoConfig(quality=20, fps=DEFAULT_FPS,
-                    save_every=DEFAULT_SAVE_EVERY)
+        stv_config.VideoConfig(quality=20, fps=DEFAULT_FPS,
+                               save_every=DEFAULT_SAVE_EVERY)
     assert "quality" in str(exc_info.value)
 
 
 def test_optimization_config_negative_steps():
     """Ensure invalid steps raises ValidationError."""
     with pytest.raises(ValidationError) as exc_info:
-        OptimizationConfig(steps=-1, style_w=DEFAULT_STYLE_WEIGHT,
-                           content_w=DEFAULT_CONTENT_WEIGHT,
-                           lr=DEFAULT_LEARNING_RATE, seed=DEFAULT_SEED,
-                           init_method=DEFAULT_INIT_METHOD)
+        stv_config.OptimizationConfig(steps=-1, style_w=DEFAULT_STYLE_WEIGHT,
+                                      content_w=DEFAULT_CONTENT_WEIGHT,
+                                      lr=DEFAULT_LEARNING_RATE,
+                                      seed=DEFAULT_SEED,
+                                      init_method=DEFAULT_INIT_METHOD)
     assert "steps" in str(exc_info.value)
 
 
 def test_optimization_config_invalid_lr():
     """Ensure invalid learning rate raises ValidationError."""
     with pytest.raises(ValidationError) as exc_info:
-        OptimizationConfig(lr=0, steps=DEFAULT_STEPS,
-                           style_w=DEFAULT_STYLE_WEIGHT,
-                           content_w=DEFAULT_CONTENT_WEIGHT,
-                           seed=DEFAULT_SEED, init_method=DEFAULT_INIT_METHOD)
+        stv_config.OptimizationConfig(lr=0, steps=DEFAULT_STEPS,
+                                      style_w=DEFAULT_STYLE_WEIGHT,
+                                      content_w=DEFAULT_CONTENT_WEIGHT,
+                                      seed=DEFAULT_SEED,
+                                      init_method=DEFAULT_INIT_METHOD)
     assert "lr" in str(exc_info.value)
 
 
 def test_optimization_config_negative_style_w():
     """Ensure invalid style weight raises ValidationError."""
     with pytest.raises(ValidationError) as exc_info:
-        OptimizationConfig(style_w=-100, steps=DEFAULT_STEPS,
-                           content_w=DEFAULT_CONTENT_WEIGHT,
-                           lr=DEFAULT_LEARNING_RATE, seed=DEFAULT_SEED,
-                           init_method=DEFAULT_INIT_METHOD)
+        stv_config.OptimizationConfig(style_w=-100, steps=DEFAULT_STEPS,
+                                      content_w=DEFAULT_CONTENT_WEIGHT,
+                                      lr=DEFAULT_LEARNING_RATE,
+                                      seed=DEFAULT_SEED,
+                                      init_method=DEFAULT_INIT_METHOD)
     assert "style_w" in str(exc_info.value)
 
 
 def test_optimization_config_negative_content_w():
     """Ensure invalid content weight raises ValidationError."""
     with pytest.raises(ValidationError) as exc_info:
-        OptimizationConfig(content_w=-5.0, steps=DEFAULT_STEPS,
-                           style_w=DEFAULT_STYLE_WEIGHT,
-                           lr=DEFAULT_LEARNING_RATE, seed=DEFAULT_SEED,
-                           init_method=DEFAULT_INIT_METHOD)
+        stv_config.OptimizationConfig(content_w=-5.0, steps=DEFAULT_STEPS,
+                                      style_w=DEFAULT_STYLE_WEIGHT,
+                                      lr=DEFAULT_LEARNING_RATE,
+                                      seed=DEFAULT_SEED,
+                                      init_method=DEFAULT_INIT_METHOD)
     assert "content_w" in str(exc_info.value)
 
 
 def test_optimization_config_negative_seed():
     """Ensure invalid seed raises ValidationError."""
     with pytest.raises(ValidationError) as exc_info:
-        OptimizationConfig(seed=-42, steps=DEFAULT_STEPS,
-                           style_w=DEFAULT_STYLE_WEIGHT,
-                           content_w=DEFAULT_CONTENT_WEIGHT,
-                           lr=DEFAULT_LEARNING_RATE,
-                           init_method=DEFAULT_INIT_METHOD)
+        stv_config.OptimizationConfig(seed=-42, steps=DEFAULT_STEPS,
+                                      style_w=DEFAULT_STYLE_WEIGHT,
+                                      content_w=DEFAULT_CONTENT_WEIGHT,
+                                      lr=DEFAULT_LEARNING_RATE,
+                                      init_method=DEFAULT_INIT_METHOD)
     assert "seed" in str(exc_info.value)
